+++ CHANGELOG.md
@@ -1,20 +1,24 @@
 # Version 0.4a-dev
 
-23/10/17
+2023/10/17
 Bug corrections.
 
 ## New functions: 
-23/10/17
+2023/10/17
 - *misc.load_ser*
 - *processing.stack_fids*
 - *Pseudo_2D.mount*
 - *Pseudo_2D.cal*
+2023/10/23
+- *fit.calc_R2*
 
 ## Modified functions:
-23/10/09
+2023/10/09
 - *misc.polyn*: use the Vandermonde matrix to compute the polynomion instead of a for loop
 23/10/17
 - *Pseudo_2D.__init__*: self.fid = None by default, added acqus parameter "TD1", removed procs parameter "cal\_1", renamed procs parameter "cal\_2" into "cal" 
+2023/10/23
+- *fit.LR*: now treated as a polynomial fit, added option to force the intercept
 
 # Version 0.4a.0